start refactoring reports to remove code duplicates
